serialization: remove windows-specific layers+base rootfs
The `rootfs.type` field previously allowed a `type` of `layers+base` to
signify that there the layers must be applied to a base image. This was
required to support windows containers in Docker initially but has since
been replaced with a more compatible mechanism. It has no relevence for
OCI.

Signed-off-by: Stephen J Day <stephen.day@docker.com>