serialization: Restore old union/layer note for layer entry
This had been part of the changeset entry, which was removed in the
last commit.  This information is already covered by the copy-on-write
note in layer.md:

  NOTE: a copy-on-write or union filesystem can make this very
  efficient:

and I'd rather stay DRY and only have this information in one place
(this entry already links to layer.md).

I also don't think the note here is an important part of *defining*
the "layer" term.

But Stephen wants it restored to provide context [1], so here it is.

[1]: https://github.com/opencontainers/image-spec/pull/243#discussion_r77416399

Signed-off-by: W. Trevor King <wking@tremily.us>