config|manifest-list: Punt to runtime-spec for platform OS/arch
The new wording makes it clear that implementations are free to
support only a subset of OSes or architectures, and both
implementations and blobs are free to support/use OSes and
architectures not listed in runtime-spec.  Leaning on runtime-spec
here lets us consolidate any OCI GOOS/GOARCH extension in a single
location (runtime-spec).

I think it is unfortunate that the manifest-list and config types have
inconsistent platform formats.  It would be nice to give config the
same 'platform' nesting as manifest-list, and to have both locations
support os.version and the other settings which are currently only in
manifest-list (or to drop those settings).  But I'll leave fixing that
for later work ;).

Signed-off-by: W. Trevor King <wking@tremily.us>