config: Remove redundant OPTIONAL from Labels description
While the other annotations references (e.g. for manifest annotations)
use the "This OPTIONAL property..." pattern, config.md puts the
OPTIONAL/REQUIRED-ness in the same line as the property name (while
manifest.md does not).  My preferred larger fix would be to follow
runtime-spec and consistently declare OPTIONAL/REQUIRED-ness in the
same line as the property name, but that's a bigger change.  This
commit just removes the redundant OPTIONAL from the description body.

Signed-off-by: W. Trevor King <wking@tremily.us>