descriptor: make definition of digest consistent
After some changes to the schema to open up the character set and add
separators to the digest algorithm, this change set ensures we have a
consistent definition for the components of a digest. The specification
has been updated to clarify this decision as well as ensure the
specification matches the validation components across the board.

The portion of a digest known as `hex` is now known as `encoded` to
correspond with the wider character set allowed.

Signed-off-by: Stephen J Day <stephen.day@docker.com>